Changed std::format to fmt::format. Moved world_mapper_builder to new file.
