+++ Vic3ToHoI4lib.vcxproj.filters
@@ -570,6 +570,9 @@
     <ClInclude Include="src\hoi4_world\world\hoi4_world_framework.h">
       <Filter>src\hoi4_world\world</Filter>
     </ClInclude>
+    <ClInclude Include="src\mappers\world\world_mapper_builder.h">
+      <Filter>src\mappers\world</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\vic3_to_hoi4_converter.cpp">
@@ -871,6 +874,9 @@
     </ClCompile>
     <ClCompile Include="src\vic3_world\world\vic3_world.cpp">
       <Filter>src\vic3_world\world</Filter>
+    </ClCompile>
+    <ClCompile Include="src\mappers\world\world_mapper_builder.cpp">
+      <Filter>src\mappers\world</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
