+++ src/vic3_world/world/vic3_world.cpp
@@ -1,6 +1,6 @@
 #include "src/vic3_world/world/vic3_world.h"
 
-#include <format>
+#include "external/fmt/include/fmt/format.h"
 
 namespace vic3
 {
@@ -40,7 +40,8 @@ WorldBuilder& WorldBuilder::AddTestStates(const std::vector<std::set<int>>& prov
    for (const auto& provinceList: provinceListLists)
    {
       const int stateNum = state_number_++;
-      this->world_options_.states.emplace(stateNum, vic3::State(vic3::StateOptions{.owner_number = stateNum, .provinces = provinceList}));
+      this->world_options_.states.emplace(stateNum,
+          vic3::State(vic3::StateOptions{.owner_number = stateNum, .provinces = provinceList}));
    }
    return *this;
 }
@@ -61,10 +62,10 @@ WorldBuilder& WorldBuilder::AddTestStateRegions(const std::vector<std::set<int>>
       std::set<std::string> provinceNames = {};
       for (auto& provinceNum: provinceSet)
       {
-         provinceNames.emplace(std::format("0x0000{:0>2}", provinceNum));
+         provinceNames.emplace(fmt::format("0x0000{:0>2}", provinceNum));
       }
 
-      this->world_options_.state_regions.emplace(std::format("REGION_{:0>3}", this->state_region_number_++),
+      this->world_options_.state_regions.emplace(fmt::format("REGION_{:0>3}", this->state_region_number_++),
           StateRegion({}, provinceNames));
    }
    return *this;
